code/telemedicine/dosewatch/app/requirements.txt
Flask==2.1.2
psycopg2-binary==2.9.5
